revised counter to size_t

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: tsuchen <tsuchen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/16 09:39:18 by tsuchen           #+#    #+#             */
-/*   Updated: 2024/05/21 22:46:40 by tsuchen          ###   ########.fr       */
+/*   Updated: 2024/05/23 23:14:27 by tsuchen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
